Move FileInfoHash to proto, synchonize implementation with peernode.
