+++ market/src/lib.rs
@@ -16,7 +16,7 @@ pub use libp2p::{
     multiaddr::{multiaddr, Protocol},
     Multiaddr,
 };
-pub use lmm::{FileInfoHash, FileResponse, SupplierInfo};
+pub use lmm::{FileResponse, SupplierInfo};
 
 pub(crate) mod behaviour;
 pub(crate) mod command;
